requirements.txt
aiohttp==3.6.2
async-timeout==3.0.1
attrs==19.3.0
certifi==2019.11.28
chardet==3.0.4
dropbox==9.4.0
emoji==0.5.4
idna==2.8
lxml==4.4.2
multidict==4.7.3
numpy==1.18.1
pandas==0.25.3
python-dateutil==2.8.1
pytz==2019.3
requests==2.22.0
six==1.13.0
slackclient==2.5.0
urllib3==1.25.7
xlrd==1.2.0
yarl==1.4.2
